packages/forge-modules/ai-chat/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "forge-modules-ai-chat"
version = "0.1.0"
description = "AI Chat module for Forge — skill-learning conversational AI with Train/Ask modes"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
dependencies = [
    "forge-framework>=0.1.0",
    "pyyaml>=6.0",
    # LLM providers — at least one must be installed at runtime
    # pip install forge-modules-ai-chat[anthropic]
    # pip install forge-modules-ai-chat[openai]
    # pip install forge-modules-ai-chat[all]
]

[project.optional-dependencies]
anthropic = ["anthropic>=0.25.0"]
openai    = ["openai>=1.0.0"]
all       = ["anthropic>=0.25.0", "openai>=1.0.0"]
